+++ .github/workflows/build-packages.yml
@@ -152,7 +152,7 @@ jobs:
 
   build-apk:
     runs-on: ubuntu-latest
-    container: golang:alpine
+    container: golang:1.22-alpine3.19
     steps:
 
       - name: Set environment
@@ -193,7 +193,7 @@ jobs:
 
   build-openapi-specs:
     runs-on: ubuntu-latest
-    container: golang:alpine
+    container: golang:1.22-alpine3.19
     steps:
         - name: Set environment
           run: |
@@ -206,8 +206,8 @@ jobs:
 
         - name: Build single OpenAPI 3.0 spec file for all documented APIs
           run: |
-            go run ./cmd/goProbe/goProbe.go --config .buildconfig/goprobe.yaml --openapi ./goprobe_${{ env.RELEASE_VERSION }}_openapi.yaml
-            go run ./cmd/global-query/main.go --config .buildconfig/global-query.yaml --openapi ./global-query_${{ env.RELEASE_VERSION }}_openapi.yaml
+            go run ./cmd/goProbe/goProbe.go -config .buildconfig/goprobe.yaml -openapi.spec-outfile ./goprobe_${{ env.RELEASE_VERSION }}_openapi.yaml
+            go run ./cmd/global-query/main.go --config .buildconfig/global-query.yaml --openapi.spec-outfile ./global-query_${{ env.RELEASE_VERSION }}_openapi.yaml
 
         - name: Store artifacts
           uses: actions/upload-artifact@v3
