+++ pkg/api/query.go
@@ -2,21 +2,11 @@ package api
 
 import (
 	"context"
-	"net/http"
 
-	"github.com/danielgtaylor/huma/v2"
 	"github.com/els0r/goProbe/pkg/query"
 	"github.com/els0r/goProbe/pkg/results"
 	"github.com/els0r/telemetry/logging"
-	"github.com/gin-gonic/gin"
 )
-
-var queryTags = []string{"Query"}
-
-// LogAndAbort logs an error and the aborts further processing
-func LogAndAbort(ctx context.Context, c *gin.Context, code int, err error) {
-	logging.FromContext(ctx).Error(c.AbortWithError(code, err))
-}
 
 func getBodyQueryRunnerHandler(caller string, querier query.Runner) func(context.Context, *ArgsInput) (*QueryResultOutput, error) {
 	return func(ctx context.Context, input *ArgsInput) (*QueryResultOutput, error) {
@@ -57,63 +47,6 @@ func runQuery(ctx context.Context, caller string, args *query.Args, querier quer
 	}
 
 	return result, nil
-}
-
-// RegisterQueryAPI registers all query related endpoints
-func RegisterQueryAPI(a huma.API, caller string, querier query.Runner, middlewares huma.Middlewares) {
-	// validation
-	huma.Register(a,
-		huma.Operation{
-			OperationID: "query-post-validate",
-			Method:      http.MethodPost,
-			Path:        ValidationRoute,
-			Summary:     "Validate query parameters",
-			Description: "Validates query parameters (1) for integrity (2) attempting to prepare a query statement from them",
-			Tags:        queryTags,
-		},
-		getBodyValidationHandler(),
-	)
-	huma.Register(a,
-		huma.Operation{
-			OperationID: "query-get-validate",
-			Method:      http.MethodGet,
-			Summary:     "Validate query parameters",
-			Path:        ValidationRoute,
-			Description: "Validates query parameters (1) for integrity (2) attempting to prepare a query statement from them",
-			Tags:        queryTags,
-		},
-		getParamsValidationHandler(),
-	)
-	// query running
-	huma.Register(a,
-		huma.Operation{
-			OperationID: "query-post-run",
-			Method:      http.MethodPost,
-			Path:        QueryRoute,
-			Summary:     "Run query",
-			Description: "Runs a query based on the parameters provided in the body",
-			Middlewares: middlewares,
-			Tags:        queryTags,
-		},
-		getBodyQueryRunnerHandler(caller, querier),
-	)
-}
-
-// ArgsBodyInput stores the query args to be validated in the body
-type ArgsInput struct {
-	Body *query.Args
-}
-
-// ArgsParamsInput stores the query args to be validated in the query parameters
-type ArgsParamsInput struct {
-	// for get parameters
-	query.Args
-	query.DNSResolution
-}
-
-// QueryResultOutput stores the result of a query
-type QueryResultOutput struct {
-	Body *results.Result
 }
 
 // getBodyValidationHandler returns the query args validation handler
